requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bleach==6.1.0
bokeh==3.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
numpy==2.0.0
openai==1.35.3
packaging==24.1
pandas==2.2.2
panel==1.4.4
param==2.1.1
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyviz_comms==3.0.2
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
webencodings==0.5.1
xyzservices==2024.6.0
